+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/item/ItemMarshmallowStick.java
@@ -1,0 +1,23 @@
+package com.codetaylor.mc.pyrotech.modules.tech.basic.item;
+
+import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasicConfig;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.init.MobEffects;
+import net.minecraft.item.ItemFood;
+import net.minecraft.item.ItemStack;
+import net.minecraft.potion.PotionEffect;
+import net.minecraft.world.World;
+
+import javax.annotation.Nonnull;
+
+public class ItemMarshmallowStick
+    extends ItemFood {
+
+  public static final String NAME = "marshmallow_stick";
+
+  public ItemMarshmallowStick() {
+
+    super(ModuleTechBasicConfig.CAMPFIRE_MARSHMALLOWS.MARSHMALLOW_HUNGER, (float) ModuleTechBasicConfig.CAMPFIRE_MARSHMALLOWS.MARSHMALLOW_SATURATION, false);
+  }
+
+}